reorganize the app
