+++ ui/app/src/project/dashboard/components/AlertErrorBoundary.tsx
