+++ ui/app/src/project/dashboard/components/ContentRefResolver.tsx
