+++ ui/app/src/project/dashboard/components/Panel.tsx
