+++ ui/app/src/project/dashboard/context/PanelContextProvider.tsx
