+++ ui/app/src/project/dashboard/context/plugin-registry/PluginBoundary.tsx
