+++ ui/app/src/project/dashboard/context/plugin-registry/PluginLoader.tsx
