+++ ui/app/src/project/dashboard/context/plugin-registry/PluginLoadingBoundary.tsx
