+++ ui/app/src/project/dashboard/context/plugin-registry/PluginRegistry.tsx
