+++ ui/app/src/project/dashboard/context/plugin-registry/bundled-plugins.ts
