+++ ui/app/src/project/dashboard/context/plugin-registry/create-plugin.tsx
