+++ ui/app/src/project/dashboard/context/plugin-registry/index.ts
