+++ ui/app/src/project/dashboard/context/plugin-registry/registry-state.ts
