+++ ui/app/src/project/dashboard/context/plugin-registry/use-plugins.ts
