+++ ui/app/src/shared/utils/url-builder.ts
